+++ components/layout/Layout.tsx
@@ -4,6 +4,8 @@ import HeadComponent from "@/components/layout/Head.tsx";
 import { TwClass } from "@/deps/styles.ts";
 import { Head } from "$fresh/src/runtime/head.ts";
 import NavBtn from "@/islands/NavBtn.tsx";
+import MainNav from "@/components/nav/MainNav.tsx";
+import ArticleNav from "@/components/nav/ArticleNav.tsx";
 
 export default function Layout({
   route,
@@ -22,6 +24,8 @@ export default function Layout({
       </Head>
       <NavBtn route={route} />
       <main class={TwClass(["flex", "justify-center"])}>
+        <MainNav route={route} />
+        <ArticleNav route={route} />
         <Article
           title={title.toLowerCase()}
           children={children}
