+++ deps/nav.ts
@@ -1,0 +1,22 @@
+export const mainNav = [
+  { name: "home", href: "/" },
+  {
+    name: "about",
+    href: "/about",
+  },
+  {
+    name: "gfx",
+    href: "/gfx",
+  },
+  {
+    name: "articles",
+    href: "/articles",
+  },
+];
+
+export const articleNav = [
+  {
+    name: "top",
+    href: "#",
+  },
+];
