+++ static/gfx-modules/model-import/index.html
@@ -16,6 +16,11 @@
       max-height: 500px;
       max-width: 250px;
     }
+    canvas {
+      border-radius: 0.25rem;
+      border-width: 1px;
+      border-color: rgba(63, 63, 70);
+    }
   </style>
   <script type="module">
     import Engine from "./target.js";
